moved sitecaptain_delete to where it was expected
